--- a/package.json
+++ b/package.json
@@ -78,7 +78,6 @@
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
     "styled-components": "^4.4.1",
-    "ts-jest": "^25.5.1",
     "typescript": "^3.8.3",
     "webpack": "^4.43.0"
   },