--- a/packages/components-providers/package.json
+++ b/packages/components-providers/package.json
@@ -10,7 +10,7 @@
   "repository": {
     "type": "git",
     "url": "https://github.com/looker-open-source/components",
-    "directory": "packages/providers"
+    "directory": "packages/components-providers"
   },
   "publishConfig": {
     "access": "public"