--- a/packages/components-test-utils/package.json
+++ b/packages/components-test-utils/package.json
@@ -10,7 +10,7 @@
   "repository": {
     "type": "git",
     "url": "https://github.com/looker-open-source/components",
-    "directory": "packages/test-utils"
+    "directory": "packages/components-test-utils"
   },
   "publishConfig": {
     "access": "public"