--- a/tsconfig.common.json
+++ b/tsconfig.common.json
@@ -27,6 +27,6 @@
     "skipLibCheck": true,
     "suppressImplicitAnyIndexErrors": true
   },
-  "exclude": ["lib", "node_modules", "**/*.test.*"],
+  "exclude": ["lib", "node_modules"],
   "include": []
 }