--- a/tsconfig.json
+++ b/tsconfig.json
@@ -5,11 +5,7 @@
     "pretty": true,
     "baseUrl": ".",
     "paths": {
-      "@looker/components": ["packages/components/src"],
-      "@looker/components-providers": ["packages/providers/src"],
-      "@looker/components-test-utils": ["packages/test-utils/src"],
-      "@looker/design-tokens": ["packages/design-tokens/src"],
-      "@looker/icons": ["packages/icons/src"]
+      "@looker/*": ["packages/*/src"]
     }
   },
   "include": ["packages/*/src"],