### :sparkles: Changes

- Jest no longer requires `yarn build` to be run beforehand. Instead, Jest now directly compiles typescript as part of the test process. 
- Switch to using dynamic package paths for `tsconfig` & Jest package resolution
- Rename `test-utils` and `providers` folders to match package names to allow for dynamic resolution
- Travis CI change: Build & test can now be run in parallel since Jest no longer uses build artifacts
- Snapshots changed shape slightly as they’re now benefitting from being generated via the Babel pipeline and the associated `babel-plugin-styled-components` plugin optimizations.

#### Benchmarks

**NOTE:** Before this change the developer was required to run `yarn build` before running `jest` to ensure that the build artifacts were in place and consistent with the source.

| Task  | Before | After |  |
|-------|-------:|------:|---------|
| `test` | ~50s + 33.8s | 34.1s | YMMV depending on cache status |
| `test --no-cache` |  ~50s + 428.9s | 81.1s |  |

Compare to: https://github.com/looker-open-source/components/pull/841


### :white_check_mark: Requirements

- [x] Includes test coverage for all changes
- [x] Build and tests are passing
- [ ] Update documentation
- [x] Updated CHANGELOG
- [ ] Checked for i18n impacts
- [ ] Checked for a11y impacts
- [x] PR is ideally < ~400LOC

### :camera: Screenshots